requirements_dev.txt
aiohttp==3.6.2
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
codecov==2.1.9
colorama==0.4.3
coverage==5.2.1
dependency-injector==3.37.0
Flask==1.1.2
Flask-Cors==3.0.9
fpl==0.6.25
gunicorn==20.0.4
idna==2.10
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
kaptan==0.5.12
MarkupSafe==1.1.1
more-itertools==8.5.0
multidict==4.7.6
neo4j-driver==1.7.2
neobolt==1.7.17
neomodel==3.3.2
neotime==1.7.4
packaging==20.4
pluggy==0.13.1
protobuf==3.13.0
PTable==0.9.2
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
pytest-aiohttp==0.3.0
pytest-cov==2.10.1
pytest-mock==3.3.1
pytz==2020.1
PyYAML==5.3.1
querystring-parser==1.2.4
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
yarl==1.5.1
